% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cleaned_data.R
\name{load_cleaned_data}
\alias{load_cleaned_data}
\title{Load clean dataset from package}
\usage{
load_cleaned_data(dataset_name, package_name)
}
\arguments{
\item{dataset_name}{Name of the dataset to load}

\item{package_name}{Name of the package containing the dataset}
}
\value{
Loaded dataset, or NULL if not found
}
\description{
Load clean dataset from package
}
