% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_length.R
\name{rb_length}
\alias{rb_length}
\title{Omzetten van lengtemetingen (km, m, cm en mm)}
\usage{
rb_length(value, from, to)
}
\arguments{
\item{value}{waarde die omgezet moet worden}

\item{from}{eenheid waarvan de waarde afkomstig is}

\item{to}{eenheid waarvan de waarde omgezet moet worden}
}
\value{
waarde met omgezette eenheid
}
\description{
Omzetten van lengtemetingen (km, m, cm en mm)
}
\examples{
#' x <- c(2, 3.5, 15, 230)
rb_length(x, from = "m", to = "mm")
}
