% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_mol.R
\name{rb_mol}
\alias{rb_mol}
\title{Molberekening (kg, g, mg, kmol, mol en mmol)}
\usage{
rb_mol(x, y, massa, molaire_massa)
}
\arguments{
\item{x}{gewicht van stof}

\item{y}{molaire gewicht van stof}

\item{massa}{gewicht eenheid (kg, g of mg)}

\item{molaire_massa}{molaire massa eenheid (kmol, mol of mmol)}
}
\value{
aantal mol van de stof
}
\description{
Molberekening (kg, g, mg, kmol, mol en mmol)
}
\examples{
x <- 32
y <- 16
rb_mol(x, y, massa = "g", molaire_massa = "mol")
}
