% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_temperature.R
\name{rb_temperature}
\alias{rb_temperature}
\title{Omzetten van temperatuur (Celcius en Kelvin)}
\usage{
rb_temperature(value, from, to)
}
\arguments{
\item{value}{waarde die omgezet moet worden}

\item{from}{eenheid waarvan de waarde afkomstig is (C of K)}

\item{to}{eenheid waarvan de waarde omgezet moet worden (C of K)}
}
\value{
waarde met omgezette eenheid
}
\description{
Omzetten van temperatuur (Celcius en Kelvin)
}
\examples{
x <- c(2, 3.5, 15, 230)
rb_temperature(x, from = "C", to = "K")
}
