% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_weight.R
\name{rb_weight}
\alias{rb_weight}
\title{Omzetten van gewichtsmetingen (kg, g en mg)}
\usage{
rb_weight(value, from, to)
}
\arguments{
\item{value}{waarde die omgezet moet worden}

\item{from}{eenheid waarvan de waarde afkomstig is}

\item{to}{eenheid waarvan de waarde omgezet moet worden}
}
\value{
waarde met omgezette eenheid
}
\description{
Omzetten van gewichtsmetingen (kg, g en mg)
}
\examples{
x <- c(2, 3.5, 15, 230)
rb_weight(x, from = "g", to = "mg")
}
